requirements.txt
python-pptx
langchain-openai
fastapi
python-multipart